delete unused assets and components, order project folders, create story section, fix navbar for mobile
