+++ public/CNAME
@@ -1,0 +1,1 @@
+nicoletologa.com
