--- src/components/LegendSection.tsx
@@ -1,10 +1,0 @@
-export default function LegendSection(){
-    return (
-    <div className="text-white py-3 bg-[#e682b6]">
-          <p className="text-xl p-6">
-            Si estas aquí es porque <strong>te sientes frustrado o has llegado a normalizar las conductas no deseadas</strong> de tu gato.
-            Como buen padre gatuno has intentado de todo, invirtiendo <strong>tiempo, dinero y esfuerzo</strong> sin obtener resultados. Puedes respirar profundo, porque no estas sol@.
-          </p>
-    </div>
-    );
-}
