+++ src/pages/EbookDetail.tsx
@@ -1,9 +1,9 @@
 import { useParams, Navigate} from 'react-router-dom';
 import { Book, ShoppingCart } from 'lucide-react';
-import Navbar from '../components/Navbar';
+import Navbar from '../components/layouts/Navbar';
 import { ebooks } from '../data/ebooks';
-import PriceDisplay from '../components/PriceDisplay';
-import Footer from '../components/Footer';
+import PriceDisplay from '../components/common/PriceDisplay';
+import Footer from '../components/layouts/Footer';
 
 export default function EbookDetail() {
   const { id } = useParams();
